feat(statements): Complete Sprint 4 - Statement Import & Fingerprinting
Implement statement import functionality with tiered fingerprinting for
automatic bank/format detection:

Backend:
- Add StatementsController with analyze/import/fingerprints endpoints
- Add TransactionsController for transaction listing and details
- Implement StatementParsingService for CSV parsing
- Implement ColumnMappingInferenceService for AI-assisted mapping (Tier 3)
- Implement StatementFingerprintService with header hash matching (Tier 1)
- Add AnalysisSessionCache for caching parsed statement data
- Add Transaction and StatementImport entities with repositories
- Add DevAuthenticationHandler for local development testing
- Fix ReceiptConfiguration to use snake_case column names

Frontend:
- Add StatementUpload component for file upload
- Add ColumnMappingEditor for column mapping configuration
- Add ImportSummary for displaying import results
- Add statementService for API integration

Testing:
- Verified Tier 1 fingerprint matching for Chase and Amex
- Verified transaction import with duplicate detection
- Verified manual mapping flow for unknown banks

+++ backend/src/ExpenseFlow.Infrastructure/Services/ThumbnailService.cs
@@ -61,18 +61,18 @@ public class ThumbnailService : IThumbnailService
         using (image)
         {
             // Resize maintaining aspect ratio, then crop to exact size
-            image.Thumbnail(new MagickGeometry(width, height)
+            image.Thumbnail(new MagickGeometry((uint)width, (uint)height)
             {
                 IgnoreAspectRatio = false,
                 FillArea = true
             });
 
             // Center crop to exact dimensions
-            if (image.Width > width || image.Height > height)
+            if (image.Width > (uint)width || image.Height > (uint)height)
             {
-                var xOffset = (image.Width - width) / 2;
-                var yOffset = (image.Height - height) / 2;
-                image.Crop(new MagickGeometry(xOffset, yOffset, width, height));
+                var xOffset = (int)((image.Width - (uint)width) / 2);
+                var yOffset = (int)((image.Height - (uint)height) / 2);
+                image.Crop(new MagickGeometry(xOffset, yOffset, (uint)width, (uint)height));
             }
 
             // Convert to JPEG
